claude-agent-lab/requirements.txt
anthropic
python-dotenv
termcolor
